test: consolidate import smokes into test_imports
